+++ tests/unit/test_sample.py
@@ -5,8 +5,10 @@ Unit tests for resdk/resources/sample.py file.
 
 import unittest
 
-from mock import MagicMock
+from mock import MagicMock, patch
 
+from resdk.exceptions import ResolweServerError
+from resdk.resources.annotations import AnnotationValue
 from resdk.resources.descriptor import DescriptorSchema
 from resdk.resources.sample import Sample
 
@@ -65,6 +67,53 @@ class TestSample(unittest.TestCase):
         sample.id = None
         with self.assertRaises(ValueError):
             _ = sample.data
+
+    def test_set_annotation(self):
+        resolwe = MagicMock()
+        # Set annotation value for existing annotation field.
+        with patch("resdk.resources.sample.Sample.annotations") as mock_annotations:
+            sample = Sample(id=1, resolwe=resolwe)
+            full_path = "general.species"
+            post_mock = MagicMock()
+            sample.api = MagicMock(return_value=post_mock)
+            annotation_value = MagicMock(spec=AnnotationValue)
+            mock_annotations.get.return_value = annotation_value
+            sample.set_annotation(full_path, "Mus musculus")
+            sample.annotations.get.assert_called_with(
+                field__name="species", field__group__name="general"
+            )
+            self.assertEqual(annotation_value.value, "Mus musculus")
+            annotation_value.save.asssert_called_once()
+
+        # Delete annotation value.
+        with patch("resdk.resources.sample.Sample.annotations") as mock_annotations:
+            sample = Sample(id=1, resolwe=resolwe)
+            full_path = "general.species"
+            post_mock = MagicMock()
+            sample.api = MagicMock(return_value=post_mock)
+            annotation_value = MagicMock(spec=AnnotationValue)
+            mock_annotations.get.return_value = annotation_value
+            sample.set_annotation(full_path, None)
+            sample.annotations.get.assert_called_with(
+                field__name="species", field__group__name="general"
+            )
+            annotation_value.delete.asssert_called_once()
+
+        # Nonexisting field.
+        resolwe.annotation_field.from_path.side_effect = LookupError
+        with patch("resdk.resources.sample.Sample.annotations") as mock_annotations:
+            sample = Sample(id=1, resolwe=resolwe)
+            full_path = "general.species"
+            post_mock = MagicMock()
+            sample.api = MagicMock(return_value=post_mock)
+            mock_annotations.get.side_effect = LookupError
+            self.assertRaisesRegex(
+                ResolweServerError,
+                "Field 'general.species' does not exist.",
+                sample.set_annotation,
+                full_path,
+                "Mus musculus",
+            )
 
     def test_set_annotations(self):
         sample = Sample(id=1, resolwe=MagicMock())